// safe-20220117/src/models/ChangeCheckRequestInstance.ts
// This file is auto-generated, don't edit it
import * as $dara from '@darabonba/typescript';


export class ChangeCheckRequestInstance extends $dara.Model {
  attributionApp?: string[];
  influenceApp?: string[];
  instance?: string[];
  nc?: string[];
  uids?: string[];
  static names(): { [key: string]: string } {
    return {
      attributionApp: 'AttributionApp',
      influenceApp: 'InfluenceApp',
      instance: 'Instance',
      nc: 'Nc',
      uids: 'Uids',
    };
  }

  static types(): { [key: string]: any } {
    return {
      attributionApp: { 'type': 'array', 'itemType': 'string' },
      influenceApp: { 'type': 'array', 'itemType': 'string' },
      instance: { 'type': 'array', 'itemType': 'string' },
      nc: { 'type': 'array', 'itemType': 'string' },
      uids: { 'type': 'array', 'itemType': 'string' },
    };
  }

  validate() {
    if(Array.isArray(this.attributionApp)) {
      $dara.Model.validateArray(this.attributionApp);
    }
    if(Array.isArray(this.influenceApp)) {
      $dara.Model.validateArray(this.influenceApp);
    }
    if(Array.isArray(this.instance)) {
      $dara.Model.validateArray(this.instance);
    }
    if(Array.isArray(this.nc)) {
      $dara.Model.validateArray(this.nc);
    }
    if(Array.isArray(this.uids)) {
      $dara.Model.validateArray(this.uids);
    }
    super.validate();
  }

  constructor(map?: { [key: string]: any }) {
    super(map);
  }
}